requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-magic==0.4.27
pillow==10.2.0
tensorflow==2.15.0
librosa==0.10.1
opencv-python==4.9.0.80
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.3
celery==5.3.6
redis==5.0.1
boto3==1.34.49
paramiko==3.4.0
prometheus-client==0.20.0
yara-python==4.5.0
alembic==1.13.1
python-docx==1.1.0
openpyxl==3.1.2
PyMuPDF==1.23.26
pytesseract==0.3.10